Add the geo_location field to assets
I have moved the related code out of files, into shared, so that it can
be used by both files and assets.
