+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [2.51.0] - 2022-05-31
+### Added
+- added the new geo_location field to the Asset resource
+
 ## [2.50.0] - 2022-05-27
 ### Changed
 - Geospatial: deprecate update_feature_types and add patch_feature_types
