+++ cognite/client/_version.py
@@ -1,2 +1,2 @@
-__version__ = "2.50.0"
+__version__ = "2.51.0"
 __api_subversion__ = "V20220125"
